% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{assessComps}
\alias{assessComps}
\title{Assess stability effect of each compartment.}
\usage{
assessComps(JM, method = "eigenvalue", MR = NULL, dead = NULL)
}
\arguments{
\item{JM}{(required) A square named Jacobian matrix with numeric values representing the effect of one compartment (rows)
on another compartment (columns).}

\item{method}{(required) Either "eigenvalue" (default), "scalar", or "initial".
\itemize{
\item{
The method "eigenvalue" finds asymptotic stability as the maximum real part of the eigenvalues
calculated from the Jacobian matrix.
}
\item{
The "scalar" method finds asymptotic stability as the scalar of natural mortality
rates needed to acquire a stable matrix.
}
\item{
The "initial" method finds initial stability as equation 4 from Arnoldi et al. (2016), based on half the the maximum
real part of the eigenvalues from the addition of the Jacobian matrix to its transpose.
}
}}

\item{MR}{(required if method is "scalar")
A named numeric vector containing mortality of the faunal compartments
(per unit time, t-1). The values and names must be in the same
order as the Jacobian matrix, and the values for dead compartments should be
set to NA.}
}
\value{
Returns a dataframe with absolute and relative changes in stability for removal
of each food web compartment from the Jacobian matrix.
}
\description{
This function assesses the affect on stability of removing individual compartments from
the Jacobian matrix.
}
\details{
If the change in stability (delta) is negative, the system becomes more stable if
the respective food web compartment is removed from the Jacobian matrix. The opposite
is true if delta is positive. \cr
The 'scalar' method might be somewhat slower than the default 'eigenvalue' method.
}
\references{
\itemize{
\item{
Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced carbon cycles and
the absence of a relation between ecosystem complexity and stability. Ecol. Lett. 17,
651–661. https://doi.org/10.1111/ele.12266
}
}
}
