% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{assessFeedback}
\alias{assessFeedback}
\title{Assess feedback characteristics of network}
\usage{
assessFeedback(JM, MR = NULL, compnames = NULL, findLoops = F,
  k = NULL, path = getwd(), file = "allLoops.txt", verbose = T)
}
\arguments{
\item{JM}{(required) Jacobian matrix with interaction strengths.}

\item{MR}{(optional) Natural mortality/death rates for scaling the
maximum loop weight, same unit as Jacobian matrix.}

\item{compnames}{(optional) Vector with compartment names in same order
as the Jacobian matrix. If it is not included the names of JM
are used as compartment names. If the JM is not named, the output will
simply include the index of compartments to indicate loops.}

\item{findLoops}{(optional) Default is FALSE. You need a text file with the indices
of each loop per line. If you do not have this yet, this function redirects
to a recursive depth-first-search function to find all loops, or all loops
of length k (see function \code{dfs()}.}

\item{k}{(optional) Integer. Can be used if findLoops is TRUE,
and indicates that you only want to search for loops of length k.}

\item{path}{(optional) Default is getwd(). Is the path were to look
for the text file with the indices of all loops.}

\item{file}{(optional) Default is "allLoops.txt". Is the text file
with indices of all loops. Should be present in the \code{path}.}

\item{verbose}{(optional) Default is TRUE.
Set to FALSE if you don't want messages printed.}
}
\value{
Returns a dataframe with all loops noted as
compName1->compName2->compNamek (column "loop"),
with feedbacks (column "fdb") and loop weights (column "lw") of
those loops.
}
\description{
This function can find all loops or loops of a specific length, and consequently
calculate the feedback and loop weight of each loop.
}
\details{
Natural death/mortality rates (MR) can be found as the absolute
values on the diagonal.
Diagonal values for detritus are also regarded 'death' rates, as they
represent self-dampening effects. \cr

The depth-first-search algorithm can find all loops in your network.
Please be aware that increasing the size of your network, exponentially
increases the computation time to find all loops. If you have a large
network you might want to limit your search to loops of length k = 2
or k = 3, as those are found to be the most important in determining
overall feedback and stability in your system (Neutel & Thorne 2014). \cr

Feedback of a loop is the product of all interaction strengths
in a loop (so feedback is not additative but multiplicative).

Loop weight is defined as the geometric mean of all absolute
interaction strengths in a loop of length k. It combines information on
the feedback of the loop, with natural death rates and the length of
the loop. The feedback is scaled by the natural death rates in the reference
because the stability measure is also scaled by death rates. Depending on
your stability measure you can decide to omit inclusion of death rates.
}
\references{
\itemize{
\item{
Neutel, A.M., Heesterbeek, J.A.P., Van De Koppel, J., Hoenderboom, G.,
Vos, A., Kaldeway, C., Berendse, F., De Ruiter, P.C., 2007. Reconciling
complexity with stability in naturally assembling food webs. Nature 449,
599–602. https://doi.org/10.1038/nature06154
}
\item{
Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced
carbon cycles and the absence of a relation between ecosystem
complexity and stability. Ecol. Lett. 17, 651–661.
https://doi.org/10.1111/ele.12266
}
}
}
