% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{assessLinksFixed}
\alias{assessLinksFixed}
\title{Assess stabiliy effect of a fixed link alteration}
\usage{
assessLinksFixed(JM, method = "eigenvalue", MR = NULL, dead = NULL,
  func = function(x) {     return(x * 2) })
}
\arguments{
\item{JM}{(required) A square named Jacobian matrix with numeric values representing the effect of one compartment (rows)
on another compartment (columns).}

\item{method}{(required) Either "eigenvalue" (default) or "scalar".
\itemize{
\item{
The method "eigenvalue" finds stability as the maximum real part of the eigenvalues
calculated from the Jacobian matrix.
}
\item{
The "scalar" method finds stability as the scalar of natural mortality
rates needed to acquire a stable matrix.
}
}}

\item{MR}{(required if method is "scalar")
A named numeric vector containing mortality of the faunal compartments
(per unit time, t-1). The values and names must be in the same
order as the Jacobian matrix, and the values for dead compartments should be
set to NA.}

\item{dead}{(optional if method is "scalar") Character vector with all names of detritus and nutrient
compartments (everything that is not fauna).}

\item{func}{(required) Function describing how to alter each interaction strength to
assess the effect on stability of the respective link. Default is doubling each
interaction strength.}
}
\value{
Returns a dataframe with absolute and relative changes in stability for alteration
of each interaction strength in the Jacobian matrix.
}
\description{
This function assesses the effect on stability of altering each interaction strength
in a Jacobian matrix according to a fixed function.
}
\details{
If the change in stability (delta) is negative, the system becomes more stable if
the respective food web compartment is removed from the Jacobian matrix. The opposite
is true if delta is positive. \cr
Using the method 'scalar' might be somewhat slower than the method 'eigenvalue'.
}
\references{
\itemize{
\item{
de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of Interaction
Strengths, and Stability in Real Ecosystems. Science (80-. ). 269, 1257–1260.
https://doi.org/10.1126/science.269.5228.1257
}
}
}
