% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_first_search.R
\name{dfs}
\alias{dfs}
\title{Depth-First-Search wrapper}
\usage{
dfs(AM, output = "allLoops", k = NULL, verbose = T)
}
\arguments{
\item{AM}{(required) Adjacency matrix}

\item{output}{(required) Name for textfile to store loops.
The data will be stored as "<output>.txt" if all loops are searched for, or as
"<output>_k=<k>.txt". Default is "allLoops", so the text file will be "allLoops.txt"
or "allLoops_k=<k>.txt".}

\item{k}{(optional) Default is NULL. Integer of length loop to search for.}

\item{verbose}{(optional) Default is TRUE. Set to FALSE to hide messages.}
}
\value{
Returns file name.
}
\description{
This function is a wrapper for two recursive Depth-First-Search functions.
The wrapper checks data input, redirects to the right algorithm, and checks if no
file will be overwritten.
}
\references{
Tarjan, R. (1972). Depth-First Search and Linear Graph Algorithms.
SIAM Journal on Computing, 1(2), 146–160.
https://doi.org/10.1137/0201010.
Implementation based on the answers at:
https://stackoverflow.com/questions/546655/finding-all-cycles-in-a-directed-graph.
}
