% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_first_search.R
\name{dfsk}
\alias{dfsk}
\title{Depth-First-Search loops of length k}
\usage{
dfsk(AM, node, visited, pathway, k, started, output, start)
}
\arguments{
\item{AM}{Adjacency matrix.}

\item{node}{Current node.}

\item{visited}{Logical vector which stores if a node is visited before.}

\item{pathway}{Integer vector that tracks the pathway to current node.}

\item{k}{Integer denoting the length of loops searched for.}

\item{started}{Logical vector which stores which nodes have been used
before as starting node.}

\item{output}{String with filename to store loops.}

\item{start}{Is integer with the current start node.}
}
\value{
No returned value, writes .txt file to working directory.
}
\description{
This recursive Depth-First-Search function finds all loops of length k starting at a
certain node in the foodweb and writes them to a text file.
}
\references{
\itemize{
\item{
Reference?
}
}
}
