% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{extractLIMdata}
\alias{extractLIMdata}
\title{Extract data from formatted LIM.}
\usage{
extractLIMdata(model, verbose = T)
}
\arguments{
\item{model}{(required) A named list containing elements with food web model data.
\itemize{
\item \code{LIM} (required) should contain a read-in LIM. Achieved through \code{Read(<path-to-input-file>)}
\item \code{web} (optional) can contain a named numeric vector with flow solutions.
If the LIM is not resolved, the function will use the parsiomious (least distance) solution.
\item \code{aTag} (optional) Tag assigned to the variables representing the assimilated amount of material.
Default is set to "ass". Not case sensitive.
\item \code{gTag} (optional) Tag assigned to the variables representing growth.
Default is set to "growth". Not case sensitive.
\item \code{mTag} (optional) Tag assigned to the flows or variables containing the natural, i.e.
non-predatory, mortality. Default is set to "mort". Not case sensitive.
\item \code{deadTag} (optional) Tag assigned to the dead compartments (e.g. detritus, carrion, nutrients).
Default is set to "dead". Not case sensitive.
\item \code{defTag} (optional) Tag assigned to the flows representing defecation.
Default is set to "def". Not case sensitive.
\item \code{setup} (optional) Setup(Read(lim.input)). By inclusion the function won't setup the model,
thereby potentially saving time.
\item \code{diagonal} (optional) Either a single value, a numeric vector or the string "model".
Default is an all-zero diagonal. The string "model" calculates the diagonal values from flux values.
If diagonal is set to "model" mortality must be explicity included in the LIM.
}}
}
\value{
Returns a list with flowmatrix \code{FM}, biomasses \code{BM}, conversion efficiencies
\code{CE$AE and CE$GE}, names of external compartments \code{externals}, information on dead
compartments \code{dead}, and mortality rates \code{MR}.
}
\description{
This function automatically extracts (physiological) data from a LIM that is formatted
with tags.
}
\details{
In order for this function to work, the LIM must be set-up in a specific way.
The flows and variables representing assimilation, growth, defecation, and mortality of an organism
must be named as: "compartmentTag" or "tagCompartment" (not case sensitive).
The dead compartments must include the deadTag.
For more information on setting up the LIM, please review the vignette.
}
\references{
\code{\link{LIM}} package, Soetaert & van Oevelen 2015.
}
\seealso{
\code{getJacobian}
}
