% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getCE}
\alias{getCE}
\title{Get conversion efficiencies from a LIM.}
\usage{
getCE(FM, vars, lim, aTag = NULL, gTag = NULL)
}
\arguments{
\item{FM}{(required) A flow matrix with flows from source in rows to sink in columns.}

\item{vars}{(required) A named vector with the values of variables defined in the LIM.}

\item{lim}{(required) Setup(Read(lim.input))}

\item{aTag}{(optional) Tag assigned to the variables containing the assimilated part. Default is set to
"ass". Not case sensitive.}

\item{gTag}{(optional) Tag assigned to the variables containing the growth part. Default is set to
"growth". Not case sensitive.}
}
\value{
Returns a named vector with assimilation efficiencies.
}
\description{
This function calculates assimilation and growth efficiences from a given LIM.
}
\details{
The LIM must be set up in a specific way. The assimilation efficiency is calculated as
the assimilated part (which is defined as variable in the LIM and calculated in the
function getVariables) divided by the ingestion (which is the sum of the organism's
column in the Flow Matrix FM).
by dividing the amount of assimilated material/energy
(must be defined in the original LIM model as variable) by the total ingestion of the organism.
This function calculates growth (or secondary production) efficiences by dividing the amount
of growth by the amount of assimilated material/energy
(both must be defined in the original LIM model as variable).
The LIM must be set up in a specific way. The assimilation efficiency is calculated as
the assimilated part (which is defined as variable in the LIM and calculated in the
function getVariables) divided by the ingestion (which is the sum of the organism's
column in the Flow Matrix FM).
!!! Growth efficiency BAC = growth / inflow
}
\references{
LIM package reference
}
