% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{getCw}
\alias{getCw}
\title{Get weighted connectance.}
\usage{
getCw(FM)
}
\arguments{
\item{FM}{(required) A flow matrix with flows from source in rows to sink in columns.}
}
\value{
Returns a double.
}
\description{
This function calculates weighted connectance, capturing the skewness of flows.
}
\details{
Cw can vary between 0 and 1. Larger values of Cw indicate a more even distribution
of fluxes.
}
\references{
\itemize{
\item{
van Altena, C., Hemerik, L., de Ruiter, P.C., 2016. Food web stability and weighted
connectance: the complexity-stability debate revisited. Theor. Ecol. 9, 49–58.
https://doi.org/10.1007/s12080-015-0291-7
}
}
}
