% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getDeadInfo}
\alias{getDeadInfo}
\title{Gets list with information on dead compartments.}
\usage{
getDeadInfo(dead, readLIM, web, FM = NULL, defTag = NULL, verbose = TRUE)
}
\arguments{
\item{dead}{(required) A list with the element "names" that contains a character vector with
names of all dead compartments.}

\item{readLIM}{(required) A linear inverse model read into the R environment with the R-package LIM.
Can be achieved with \code{readLIM <- Read("path_to_input_file")}}

\item{web}{(required) A named numeric vector with the flow solutions.}

\item{FM}{(optional) A flow matrix with flows from source in rows to sink in columns.}

\item{defTag}{(optional) Tag assigned to the flows representing defecation.
Default is set to "def". Not case sensitive.}

\item{verbose}{(optional) Should the function provide messages?}
}
\value{
Returns a list with two elements:
\itemize{
\item{names: a character vector with names of dead compartments}
\item{frac: a matrix the same size as FM containing the fraction of each flow that is defecation}
}
}
\description{
This function determines for each dead compartment whether or not defecation occurs into it, and,
if parallel flows occur, what fraction of each flow is defecation.
}
\details{
In order for this function to work, the LIM must be set-up in a specific way.
The flows representing defecation of an organism must be named as:
"compartmentTag" or "tagCompartment" (not case sensitive).
For more information on setting up the LIM, please review the vignette.
}
\references{
\code{\link{LIM}} package, Soetaert & van Oevelen 2015.
}
\seealso{
\code{getFlowMatrix}
}
