% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagonal.R
\name{getDiagonalSpecies}
\alias{getDiagonalSpecies}
\title{Calculate diagonal value for species}
\usage{
getDiagonalSpecies(MR, BM, index = NULL)
}
\arguments{
\item{MR}{A named numeric vector with non-predatory mortality rates for all
compartments (biomass per unit time or biomass per unit time per surface area).
(required)}

\item{BM}{A named numeric vector with biomasses of all compartments
(just biomass or biomass per unit area), must be in the same order as MR. (required)}
}
\value{
This function returns a named numeric vector with diagonal values for the
species in the food web (per unit time). It is important to review the units of the
input data. If MR is biomass per unit time then BM must be just biomass. If MR is
biomass per unit time per surface area then BM must be biomass per surface area.
}
\description{
This function finds the diagonal values for species to be used in the Jacobian matrix.
The calculation is based on their non-predatory mortality rate and biomass, and is
an implementation of the equation from Neutel & Thorne (2014). \cr
References: \cr
Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced carbon cycles
and the absence of a relation between ecosystem complexity and stability. Ecol. Lett. 17,
651–661. https://doi.org/10.1111/ele.12266
}
