% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{getFeedback}
\alias{getFeedback}
\title{Get loop feedback}
\usage{
getFeedback(IS)
}
\arguments{
\item{IS}{(required) Numeric vector with interaction strengths in a loop.}
}
\value{
Returns a double.
}
\description{
This function calculates the feedback of a loop.
}
\details{
Feedback of a loop is the product of all interaction strengths
in a loop (so feedback is not additative but multiplicative).
}
\references{
\itemize{
\item{
Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced
carbon cycles and the absence of a relation between ecosystem
complexity and stability. Ecol. Lett. 17, 651–661.
https://doi.org/10.1111/ele.12266
}
}
}
