% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getFlowMatrix}
\alias{getFlowMatrix}
\title{Get a Flowmatrix from a LIM}
\usage{
getFlowMatrix(readLIM, web = NULL, lim = NULL)
}
\arguments{
\item{readLIM}{(required) A linear inverse model read into the R environment with the R-package LIM.
Can be achieved through \code{readLIM <- Read("path_to_input_file")}}

\item{web}{(optional) The solved (food) web problem, i.e. the values of the unknowns;
if not specified, the model is solved first, using Ldei.}
}
\value{
Returns an energy-flux model in the format needed for the getJacobian function.
}
\description{
This function extracts flow data from the LIM and produces a Flowmatrix i.e. flows from sources
in rows to the sinks in columns.
}
\details{
The LIM must be set up in a specific way. sum columns must equal sum rows.
}
\references{
LIM package reference
}
