% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{getJacobianEnergyFlux}
\alias{getJacobianEnergyFlux}
\title{Jacobian matrix with interaction strengths}
\usage{
getJacobianEnergyFlux(FM, BM, AE, GE, diagonal = NULL, dead = NULL,
  externals = NULL, MR = NULL, verbose = T, netto = NULL)
}
\arguments{
\item{FM}{(required) A named square flowmatrix, source compartments as rows,
sink compartments as columns.}

\item{BM}{(required) A named numeric vector with biomasses of all compartments, must be in the same
order as the flow matrix \code{FM} after externals are excluded (see \code{externals}).}

\item{AE}{(required) A named numeric vector with assimilation efficiencies of all
compartments, must be in the same order as the flow matrix \code{FM} after externals
are excluded. \code{AE} should be set to NA for dead/non-faunal compartments
(see \code{dead}).}

\item{GE}{(required) A named numeric vector with growth efficiencies of all compartments,
must be in the same order as the flow matrix \code{FM} after externals are excluded.
\code{GE} should be set to NA for dead/non-faunal compartments (see \code{dead} and
\code{externals}).}

\item{diagonal}{(required) Either a single value, a numeric vector or the string "model".
Default is an all-zero diagonal.
\itemize{
\item A single value will set all diagonal values to this number.
\item A vector will set the diagonal to this user-specified diagonal.
\item The string "model" calculates the diagonal values from flux values.
This requires the argument \code{MR}.
}}

\item{dead}{(optional) List with at most three elements named \emph{names}, \emph{def},
and \emph{frac} containing information on all dead compartments (like detritus and nutrients).
\itemize{
\item The element \emph{names} is required and contains a character vector with all names of dead
compartments.
\item The element \emph{def} of the list is optional and can contain a character vector
specifying for each dead compartment if defecation occurs into this compartment
(\bold{Def}) or not (\bold{noDef}).
If this element is omitted it is assumed no defecation occurs into all dead compartments.
If there are multiple defecation compartments, the Flowmatrix is used to
calculate the relative distribution of matter into the specified defecation
compartments.
\item The element \emph{frac} is optional and should contain a matrix the same size and order
as the \code{FM} matrix with the fraction of each flow that is defecation.
This information is only needed if there are multiple parallel flows between two compartments of
which only one reflects actual defecation, and the rest is for example mortality.
}}

\item{externals}{(optional) Character vector with all names of external
compartments, i.e. which have no biomass, that have to be removed from
the flow matrix before calculations.}

\item{MR}{(required when \code{diagonal} is set to \emph{model})
A named numeric vector with non-predatory mortality rates for all
compartments per unit time (t-1)).}

\item{verbose}{(optional) Default is TRUE. Wether or not to print messages.}

\item{netto}{(optional) Boolean. Default is NULL. If TRUE, the netto Flowmatrix is used
to calculate interaction strengths. This is only relevant if there are two food web compartments
which act both as prey and predators to one another.}
}
\value{
This function returns a matrix containing interaction strengths, i.e. the
effect of the resources (rows) on the consumers (columns) - for all
interactions in the food web.
}
\description{
This functions calculates interaction strengths from a resolved energy-flux
food web model and uses these values as entries for a Jacobian matrix.
The diagonal can be all-zero (default), user-defined, or calculated from the model.
This function is an implementation of the equations from the references listed below.
}
\details{
\code{MR} can sometimes be extracted from the food web model, for example when
natural death results in a flux from the faunal compartment to a carcass compartment.
It can also be calculated as the inverse of the natural lifespan of the species
(per unit time) multiplied by the biomass of the compartment.
}
\references{
\itemize{
\item de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of
Interaction Strengths, and Stability in Real Ecosystems. Science. 269,
1257–1260. https://doi.org/10.1126/science.269.5228.1257
\item Neutel, A.M., Heesterbeek, J.A.P., Ruiter, P.C. De, 2002. Stability in Real Food Webs:
Weak Links in Long Loops. Science. 296, 1120–1123. https://doi.org/10.1126/science.1068326
\item Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced carbon cycles
and the absence of a relation between ecosystem complexity and stability. Ecol. Lett. 17,
651–661. https://doi.org/10.1111/ele.12266
}
}
