% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{getLoopWeight}
\alias{getLoopWeight}
\title{Get loop weight.}
\usage{
getLoopWeight(IS, d = NULL)
}
\arguments{
\item{IS}{(required) Numeric vector of length \code{k} with interaction strengths.}

\item{d}{(optional) Numeric vector of length \code{k} with natural death rates in
the same unit as IS.}
}
\value{
Returns a double.
}
\description{
This function calculates the loop weight.
}
\details{
Loop weight is defined as the geometric mean of all absolute
interaction strengths in a loop of length k. It combines information on
the feedback of the loop, with natural death rates and the length of
the loop. The feedback is scaled by the natural death rates in the reference
because the stability measure is also scaled by death rates. Depending on
your stability measure you can decide to omit inclusion of death rates.
}
\references{
\itemize{
\item{
Neutel, A.M., Heesterbeek, J.A.P., Van De Koppel, J., Hoenderboom, G.,
Vos, A., Kaldeway, C., Berendse, F., De Ruiter, P.C., 2007. Reconciling
complexity with stability in naturally assembling food webs. Nature 449,
599–602. https://doi.org/10.1038/nature06154
}
}
}
