% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diagonal.R
\name{getMortalityRates}
\alias{getMortalityRates}
\title{Calculate mortality rates from energy-fluxes}
\usage{
getMortalityRates(FM, AE, GE, BM, dead_names)
}
\arguments{
\item{FM}{(required) A named square flowmatrix, source compartments as rows,
sink compartments as columns. Should NOT contain external compartments.}

\item{AE}{(required) A named numeric vector with assimilation efficiencies of all
compartments, must be in the same order as the flow matrix after externals
are excluded. Must be a fraction i.e. between 0 and 1.}

\item{GE}{(required) A named numeric vector with growth (production) efficiencies of all compartments.
Must be a fraction i.e. between 0 and 1.}

\item{BM}{(required) A named numeric vector with biomasses of all compartments, must be in the same
order as the flow matrix after externals are excluded.}

\item{dead_names}{(optional) Character vector with all names of detritus and nutrient
compartments (everything that is not fauna).}
}
\value{
Returns a names vector with mortality rates in the unit per time, t-1.
}
\description{
This function can be used to calculate mortality rates (unit t-1) based on a Flowmatrix,
conversion efficiencies, and biomasses.
}
