% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getMR}
\alias{getMR}
\title{Get mortality rates from a LIM.}
\usage{
getMR(BM, web, mTag = NULL)
}
\arguments{
\item{BM}{(required) A named vector containing biomasses of all LIM components.}

\item{web}{(required) A named vector with the flow values.}

\item{mTag}{(optional) Tag assigned to the variables/flows containing the natural
flow mortality. Default is set to "mort". Not case sensitive.}
}
\value{
Returns a named vector with mortality rates (per unit time).
}
\description{
This function calculates the mortality rates (per unit time) of all organisms from a LIM.
}
\details{
The LIM must be set up in a specific way. The assimilation efficiency is calculated as
the assimilated part (which is defined as variable in the LIM and calculated in the
function getVariables) divided by the ingestion (which is the sum of the organism's
column in the Flow Matrix FM).
by dividing the amount of assimilated material/energy
(must be defined in the original LIM model as variable) by the total ingestion of the organism.
This function calculates growth (or secondary production) efficiences by dividing the amount
of growth by the amount of assimilated material/energy
(both must be defined in the original LIM model as variable).
The LIM must be set up in a specific way. The assimilation efficiency is calculated as
the assimilated part (which is defined as variable in the LIM and calculated in the
function getVariables) divided by the ingestion (which is the sum of the organism's
column in the Flow Matrix FM).
!!! Growth efficiency BAC = growth / inflow
}
\references{
LIM package reference
}
