% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability.R
\name{getStability}
\alias{getStability}
\title{Get the mathematical stability of a matrix.}
\usage{
getStability(JM)
}
\arguments{
\item{JM}{A Jacobian matrix.}
}
\value{
a numeric value, a negative value indicates a stable matrix.
}
\description{
This function allows you to find the stability of a state matrix by
finding the maximum value of the real part of its eigenvalues.
}
\examples{
getStability(JM)
}
