% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{assessComps}
\alias{assessComps}
\title{Assess stabiliy effect of each compartment.}
\usage{
assessComps(JM)
}
\arguments{
\item{JM}{(required) A Jacobian matrix.}
}
\value{
Returns a dataframe with absolute and relative changes in stability for removal
of each food web compartment from the Jacobian matrix.
}
\description{
This function assesses the affect on stability of removing one compartment from
the Jacobian matrix.
}
\details{
If the change in stability (delta) is negative, the system becomes more stable if
the respective food web compartment is removed from the Jacobian matrix. The opposite
is true if delta is positive.
}
\references{
\itemize{
\item{
Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced carbon cycles and
the absence of a relation between ecosystem complexity and stability. Ecol. Lett. 17,
651–661. https://doi.org/10.1111/ele.12266
}
}
}
