% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{assessLinksFixed}
\alias{assessLinksFixed}
\title{Assess stabiliy effect of a fixed link alteration}
\usage{
assessLinksFixed(JM, func = function(x) {     return(x * 2) })
}
\arguments{
\item{JM}{(required) A Jacobian matrix.}

\item{func}{(required) Function describing how to alter each interaction strength to
assess the effect on stability of the respective link. Default is doubling each
interaction strength.}
}
\value{
Returns a dataframe with absolute and relative changes in stability for alteration
of each interaction strength in the Jacobian matrix.
}
\description{
This function assesses the effect on stability of altering each interaction strength
in a Jacobian matrix according to a fixed function.
}
\details{
If the change in stability (delta) is negative, the system becomes more stable if
the respective food web compartment is removed from the Jacobian matrix. The opposite
is true if delta is positive.
}
\references{
\itemize{
\item{
de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of Interaction
Strengths, and Stability in Real Ecosystems. Science (80-. ). 269, 1257–1260.
https://doi.org/10.1126/science.269.5228.1257
}
}
}
