% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{assessLinksFixed}
\alias{assessLinksFixed}
\title{Assess stabiliy effect of a fixed link alteration}
\usage{
assessLinksFixed(
  JM,
  method = "eigenvalue",
  MR = NULL,
  dead = NULL,
  func = function(x) {     return(x * 2) }
)
}
\arguments{
\item{JM}{(required) A square named Jacobian matrix with numeric values representing the effect of one compartment (rows)
on another compartment (columns).}

\item{method}{(required) Either "eigenvalue" (default), "scalar", or "initial".
\itemize{
\item{
The method "eigenvalue" finds asymptotic stability as the maximum real part of the eigenvalues
calculated from the Jacobian matrix.
}
\item{
The "scalar" method finds asymptotic stability as the scalar of natural mortality
rates needed to acquire a stable matrix.
}
\item{
The "initial" method finds initial stability as equation 4 from Arnoldi et al. (2016);
initial stability is half the maximum real part of the eigenvalues derived from the
matrix obtained by addition of the Jacobian matrix to its transpose.
}
}}

\item{MR}{(required if method is "scalar")
A named numeric vector containing mortality of the faunal compartments
(per unit time, t-1). The values and names must be in the same
order as the Jacobian matrix, and the values for dead compartments should be
set to NA.}

\item{func}{(required) Function describing how to alter each interaction strength to
assess the effect on stability of the respective link. Default is doubling each
interaction strength (interaction strenght *2).}
}
\value{
Returns a dataframe with absolute and relative changes in stability for alteration
of each interaction strength in the Jacobian matrix.
}
\description{
This function assesses the effect of altering each interaction strength
in a Jacobian matrix according to a fixed function on stability .
}
\details{
If the change in stability (delta) is negative, the system becomes more stable if
the respective interaction strength is altered in the Jacobian matrix
(all other interaction strenghts keep their original value).
The system becomes less stable if delta is positive. \cr
Using the method 'scalar' might be somewhat slower than the method 'eigenvalue'.
}
\references{
\itemize{
\item{
de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of Interaction
Strengths, and Stability in Real Ecosystems. Science (80-. ). 269, 1257–1260.
https://doi.org/10.1126/science.269.5228.1257
}
}
}
