% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{assessLinksPerm}
\alias{assessLinksPerm}
\title{Assess probability of destabilizing system per interaction}
\usage{
assessLinksPerm(JM, method = "eigenvalue", mortalities = NULL,
  dead = NULL, perms = 100, threshold = 0.01, seed = 1)
}
\arguments{
\item{JM}{(required) A square named Jacobian matrix with numeric values representing the effect of one compartment (rows)
on another compartment (columns).}

\item{method}{(required) Either "eigenvalue" (default) or "scalar".
\itemize{
\item{
The method "eigenvalue" finds stability as the maximum real part of the eigenvalues
calculated from the Jacobian matrix.
}
\item{
The "scalar" method finds stability as the scalar of natural mortality
rates needed to acquire a stable matrix.
}
}}

\item{dead}{(optional if method is "scalar") Character vector with all names of detritus and nutrient
compartments (everything that is not fauna).}

\item{perms}{(required) Default is 100. Number of times a pair of interaction strengths
is varied.}

\item{threshold}{(required) Default is 0.01. The Jacobian matrix is set at this fraction below
the stability threshold before starting permutations.}

\item{seed}{(required) Default is 1. Set seed to allow reproducability of results.}
}
\value{
Returns a dataframe with the probability (0 to 1) of destabilizing the system for each
pair of trophic interactions.
}
\description{
This function determines the probability that a system becomes unstable if the
strengths in an interaction are varied.
}
\details{
First, the diagonal of the Jacobian matrix is altered so that the matrix is
just stable. This is done by subtracting the maximum real part of the eigenvalues from the
diagonal values, and subsequently reducing the diagonal values by a fraction
(in argument \code{threshold}). The default threshold of 0.01 is used in De Ruiter et al. (1995).
Every pair of interaction strengths is then randomly varied between 0 and 2aij a certain
number of times. The probability that the food web becomes unstable is the total count of
unstable matrices divided by the total number of permutations. \cr
Using the method 'scalar' to asses stablity might be somewhat slower than the method 'eigenvalue'.
}
\references{
\itemize{
\item{
de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of Interaction
Strengths, and Stability in Real Ecosystems. Science (80-. ). 269, 1257–1260.
https://doi.org/10.1126/science.269.5228.1257
}
}
}
\seealso{
\code{getStability}
}
