% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_first_search.R
\name{dfs}
\alias{dfs}
\title{Depth-First-Search wrapper}
\usage{
dfs(AM, k = NULL, output = NULL, verbose = T)
}
\arguments{
\item{AM}{(required) Adjacency matrix}

\item{k}{(optional) Integer of length loop to search for.}

\item{output}{(optional) Filename to store loops. Default is "allLoops.txt"
or "allLoops_k=<k>.txt".}

\item{verbose}{(optional) Default is TRUE. Set to FALSE to hide messages.}
}
\value{
Returns file name.
}
\description{
This function is a wrapper for two recursive Depth-First-Search functions.
The wrapper checks data input, redirects to the right algorithm, and checks if no
file will be overwritten.
}
\references{
\itemize{
\item{
Reference?
}
}
}
