% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_first_search.R
\name{dfsall}
\alias{dfsall}
\title{Depth-First-Search all loops}
\usage{
dfsall(AM, node, visited, pathway, started, file, start)
}
\arguments{
\item{AM}{Adjacency matrix.}

\item{node}{Current node.}

\item{visited}{Logical vector which stores if a node is visited before.}

\item{pathway}{Integer vector that tracks the pathway to current node.}

\item{started}{Logical vector which stores which nodes have been used
before as starting node.}

\item{file}{String with name for textfile that will store loops.}

\item{start}{Is integer with the current start node.}
}
\value{
No returned value, writes .txt file to working directory.
}
\description{
This recursive Depth-First-Search function finds all loops in the foodweb starting from
a certain node and writes them to a text file.
}
