% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{effectOnResource}
\alias{effectOnResource}
\title{Per capita effect on resource population.}
\usage{
effectOnResource(FM, BM, AE, dead)
}
\arguments{
\item{FM}{A square flowmatrix, source compartments as rows,
sink compartments as columns.}

\item{BM}{Numeric vector with biomasses of all compartments, must be in the same order as the flow matrix.}

\item{AE}{Numeric vector with assimilation efficiencies of all
compartments,must be in the same order as the flow matrix.}

\item{dead}{Character vector with all names of detritus and nutrient
compartments (everything that is not fauna).}
}
\value{
A matrix containing interaction strengths - measured as the
effect of the consumer (rows) on the resource (columns) - for all
interactions in the food web.
}
\description{
This function finds the per capita effect of a consumer on the growth
of the resource population. It uses functions from de Ruiter et al.
(1995), Neutel et al. (2002) and Neutel & Thorne (2014). The flow
from prey to predator is divided by the biomass of the consumer.
The flow from consumer to detritus, plus the defecation of predators of
the consumer into the detritus pool, minus the uptake of detritus, is
divided by the biomass of the consumer.
}
\examples{
effectOnResource(Flowmatrix(model), BM = c(10,20,30), AE = c(0.2, 0.6, 0.5), dead = "DETRITUS")
}
