% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{effectOnResource}
\alias{effectOnResource}
\title{Per capita effect on resource population.}
\usage{
effectOnResource(FM, BM, AE, dead = NULL)
}
\arguments{
\item{FM}{A square flowmatrix, source compartments as rows,
sink compartments as columns. (required)}

\item{BM}{Numeric vector with biomasses of all compartments,
must be in the same order as the flow matrix. (required)}

\item{AE}{Numeric vector with assimilation efficiencies of all
compartments, must be in the same order as the flow matrix. (required)}

\item{dead}{Integer vector with indices of all detritus and nutrient
compartments (everything that is not fauna). (optional)}
}
\value{
This function returns a matrix containing interaction strengths - measured as the
effect of the consumer (rows) on the resource (columns) - for all
interactions in the food web.
}
\description{
This function finds the per capita effect of a consumer on the growth
of the resource population. It implements the functions from de Ruiter et al.
(1995), Neutel et al. (2002) and Neutel & Thorne (2014). The flow
from prey to predator is divided by the biomass of the consumer.
The flow from consumer to detritus, plus the defecation of predators of
the consumer into the detritus pool, minus the uptake of detritus, is
divided by the biomass of the consumer. \cr
References: \cr
de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of
Interaction Strengths, and Stability in Real Ecosystems. Science. 269,
1257–1260. https://doi.org/10.1126/science.269.5228.1257 \cr
Neutel, A.M., Heesterbeek, J.A.P., Ruiter, P.C. De, 2002. Stability in Real Food Webs:
Weak Links in Long Loops. Science. 296, 1120–1123. https://doi.org/10.1126/science.1068326 \cr
Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced carbon cycles
and the absence of a relation between ecosystem complexity and stability. Ecol. Lett. 17,
651–661. https://doi.org/10.1111/ele.12266
}
\examples{
effectOnResource(Flowmatrix(model), BM = c(10,20,30), AE = c(0.2, 0.6, 0.5),
dead = "DETRITUS")
}
