% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{effectOnResource}
\alias{effectOnResource}
\title{Per capita effect on resource population.}
\usage{
effectOnResource(FMs, BM, AE, dead = NULL)
}
\arguments{
\item{FMs}{(required) A list with two elements \code{original} and \code{netMatrix}, both
containing a square flowmatrix, with source compartments as rows, sink compartments as columns.
\itemize{
\item{
the list element \code{original} should contain the original flow matrix as can be obtained
with the function \code{getFlowMatrix},
i.e. can contain a flow from A to B, and a flow from B to A.
}
\item{
the list element \code{netMatrix} should contain an adjusted version of the original matrix with only
net flows, which can be obtained with the function \code{getNetMatrixFM},
i.e. a netMatrix flow is the absolute result of flow A to B minus flow B to A in the correct direction.
The net flow is only calculated for flows between live compartments, not for dead ones.
}
}}

\item{BM}{(required) Numeric vector with biomasses of all compartments,
must be in the same order as the flow matrix.}

\item{AE}{(required) Numeric vector with assimilation efficiencies of all
compartments, must be in the same order as the flow matrix.}

\item{dead}{(optional) List with at most two elements named \emph{names} and
\emph{frac} containing information on all dead compartments (like detritus and nutrients).
\itemize{
\item The element \emph{names} is required and contains a character vector with all names of dead
compartments.
\item The element \emph{frac} is required if there is more than one detritus compartment.
It is matrix the same size and order
as the \code{FM} matrix, and contains the fraction of each flow that is defecation.
If there are multiple defecation compartments, the Flowmatrix \code{FM} combined with \code{frac} is used to
calculate the relative distribution of matter into defecation compartments.
}}
}
\value{
This function returns a matrix containing the effects of the consumers (rows) on the
resources (columns).
}
\description{
This function finds the per capita effect of a consumer on the growth
of the resource population. It implements the functions from de Ruiter et al.
(1995), Neutel et al. (2002) and Neutel & Thorne (2014). The flow
from prey to predator is divided by the biomass of the consumer.
The flow from consumer to detritus, plus the defecation of predators of
the consumer into the detritus pool, minus the uptake of detritus, is
divided by the biomass of the consumer.
}
\references{
\itemize{
\item de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of
Interaction Strengths, and Stability in Real Ecosystems. Science. 269,
1257–1260. https://doi.org/10.1126/science.269.5228.1257
\item Neutel, A.M., Heesterbeek, J.A.P., Ruiter, P.C. De, 2002. Stability in Real Food Webs:
Weak Links in Long Loops. Science. 296, 1120–1123. https://doi.org/10.1126/science.1068326
\item Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced carbon cycles
and the absence of a relation between ecosystem complexity and stability. Ecol. Lett. 17,
651–661. https://doi.org/10.1111/ele.12266
}
}
