% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{effectOnResource}
\alias{effectOnResource}
\title{Per capita effect on resource population.}
\usage{
effectOnResource(FMs, BM, AE, dead = NULL, index = NULL)
}
\arguments{
\item{BM}{(required) Numeric vector with biomasses of all compartments,
must be in the same order as the flow matrix.}

\item{AE}{(required) Numeric vector with assimilation efficiencies of all
compartments, must be in the same order as the flow matrix.}

\item{dead}{(optional) List with at most three elements named \emph{names}, \emph{def},
and \emph{frac} containing information on all dead compartments (like detritus and nutrients).
\itemize{
\item The element \emph{names} is required and contains a character vector with all names of dead
compartments.
\item The element \emph{def} should be either NULL or contain a character vector
specifying for each dead compartment if defecation occurs into this compartment
(\bold{Def}) or not (\bold{noDef}).
If this element is NULL it is assumed no defecation occurs into all dead compartments.
If there are multiple defecation compartments, the Flowmatrix \code{FM} is used to
calculate the relative distribution of matter into the specified defecation
compartments.
\item The element \emph{frac} should be either NULL or contain a matrix the same size and order
as the \code{FM} matrix with the fraction of each flow that is defecation.
This information is only needed if there are multiple parallel flows between two compartments of
which only one reflects actual defecation, and the rest is for example mortality.
}}

\item{FM}{(required) A square flowmatrix, source compartments as rows,
sink compartments as columns.}
}
\value{
This function returns a matrix containing the effects of the consumers (rows) on the
resources (columns).
}
\description{
This function finds the per capita effect of a consumer on the growth
of the resource population. It implements the functions from de Ruiter et al.
(1995), Neutel et al. (2002) and Neutel & Thorne (2014). The flow
from prey to predator is divided by the biomass of the consumer.
The flow from consumer to detritus, plus the defecation of predators of
the consumer into the detritus pool, minus the uptake of detritus, is
divided by the biomass of the consumer.
}
\references{
\itemize{
\item de Ruiter, P.C., Neutel, A.M., Moore, J.C., 1995. Energetics, Patterns of
Interaction Strengths, and Stability in Real Ecosystems. Science. 269,
1257–1260. https://doi.org/10.1126/science.269.5228.1257
\item Neutel, A.M., Heesterbeek, J.A.P., Ruiter, P.C. De, 2002. Stability in Real Food Webs:
Weak Links in Long Loops. Science. 296, 1120–1123. https://doi.org/10.1126/science.1068326
\item Neutel, A.M., Thorne, M.A.S., 2014. Interaction strengths in balanced carbon cycles
and the absence of a relation between ecosystem complexity and stability. Ecol. Lett. 17,
651–661. https://doi.org/10.1111/ele.12266
}
}
