% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getCE}
\alias{getCE}
\title{Get conversion efficiencies from a LIM.}
\usage{
getCE(FM, vars, lim, aTag = NULL, gTag = NULL, verbose = T)
}
\arguments{
\item{FM}{(required) A flow matrix with flows from source in rows to sink in columns.}

\item{vars}{(required) A named vector with the values of variables defined in the LIM.}

\item{lim}{(required) Setup(Read(lim.input))}

\item{aTag}{(optional) Tag assigned to the variables representing the assimilated amount of material.
Default is set to "ass". Not case sensitive.}

\item{gTag}{(optional) Tag assigned to the variables representing growth.
Default is set to "growth". Not case sensitive.}
}
\value{
Returns a list with element "AE" (named vector with assimilation efficiencies) and element
"GE" (named vector with growth efficiencies).
}
\description{
This function calculates assimilation and growth efficiences from a given LIM.
}
\details{
In order for this function to work, the LIM must be set-up in a specific way.
The variables representing total assimilation and growth of an organism must be named as:
"compartmentTag" or "tagCompartment" (not case sensitive).
All non-dead organisms must have a variable with assimilation and growth.
For more information on setting up the LIM, please review the vignette.

The assimilation efficiency is calculated as amount of assimilated material divided by the
total ingestion. Growth (or secondary production) efficiency is calculated by dividing
growth by the amount of assimilated material.
}
\references{
\code{\link{LIM}} package, Soetaert & van Oevelen 2015.
}
