% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{getConnPerNode}
\alias{getConnPerNode}
\title{Get effective or topological connectance per node.}
\usage{
getConnPerNode(FM, type = c("eff", "topo"))
}
\arguments{
\item{FM}{(required) A flow matrix with flows from source in rows to sink in columns.}

\item{type}{\itemize{
\item{"eff" calculates the effective connectance per node (m).}
\item{"topo" calculates the topological connectance per node (m*), which is the special
case where all flows have equal weights.}
}}
}
\value{
Returns a double.
}
\description{
This function calculates effective or topological connectance per node, which can
plotted in the window of vitality.
}
\references{
\itemize{
\item{
van Altena, C., Hemerik, L., de Ruiter, P.C., 2016. Food web stability and weighted
connectance: the complexity-stability debate revisited. Theor. Ecol. 9, 49–58.
https://doi.org/10.1007/s12080-015-0291-7
}
\item{Ulanowicz 1997.}
}
}
