% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{getCw}
\alias{getCw}
\title{Get weighted connectance.}
\usage{
getCw(FM)
}
\arguments{
\item{FM}{(required) A flow matrix with flows from source in rows to sink in columns.}
}
\value{
Returns a double.
}
\description{
This function calculates weighted connectance, capturing system complexity and
the skewness of flows.
}
\details{
Cw can vary between 1/number of trophic groups and 1. Larger values of Cw indicate a more even distribution
of fluxes. Cw is used as a measure of system complexity.
}
\references{
\itemize{
\item{
van Altena, C., Hemerik, L., de Ruiter, P.C., 2016. Food web stability and weighted
connectance: the complexity-stability debate revisited. Theor. Ecol. 9, 49–58.
https://doi.org/10.1007/s12080-015-0291-7
}
\item{
Boit, A.; Gaedke, U., 2014. Benchmarking successional progress in a quantitative food web.
Plos One 9(2):e90404. doi:10.1371/journal.pone.0090404
}
}
}
