% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_characteristics.R
\name{getCw}
\alias{getCw}
\title{Get weighted connectance.}
\usage{
getCw(FM, cannibalism = FALSE)
}
\arguments{
\item{FM}{(required) A flow matrix with flows from source in rows to sink in columns.}

\item{cannibalism}{(optional) If cannibalism occurs in the food web.}
}
\value{
Returns a double.
}
\description{
This function calculates weighted connectance, capturing the skewness of flows.
}
\references{
\itemize{
\item{van Altena et al. 2016
}
}
}
