% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getFlowMatrix}
\alias{getFlowMatrix}
\title{Get a Flowmatrix from a LIM}
\usage{
getFlowMatrix(originalFM = NULL, flows = NULL, answers = NULL)
}
\arguments{
\item{lim}{(required) A lim model.}
}
\value{
Returns an energy-flux model in the format needed for the getJacobian function.
}
\description{
This function extracts flow data from the LIM and produces a Flowmatrix i.e. flows from sources
in rows to the sinks in columns.
}
\details{
The LIM must be set up in a specific way.
}
\references{
LIM package reference
}
