% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getFlowMatrix}
\alias{getFlowMatrix}
\title{Get Flowmatrix from LIM}
\usage{
getFlowMatrix(readLIM, web = NULL, lim = NULL)
}
\arguments{
\item{readLIM}{(required) A linear inverse model read into the R environment with the R-package LIM.
Can be achieved with \code{readLIM <- Read("path_to_input_file")}}

\item{web}{(optional) The solved (food) web problem, i.e. the values of the unknowns;
if not specified, the model is solved first, using Ldei.}

\item{lim}{(optional) The set-up linear problem.
Can be achieved with \code{lim <- Setup(readLIM)}.}
}
\value{
Returns a named numeric matrix.
}
\description{
This function extracts flow data from the LIM and produces a Flowmatrix i.e. flows from sources
in rows to the sinks in columns.
}
\details{
This function is very similar to the \code{Flowmatrix} function from the LIM package.
The major difference is that this function also provides the right answer if multiple parallel flows occur
between the same two compartments.
}
\references{
\code{\link{LIM}} package, Soetaert & van Oevelen 2015.
}
