% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{getJacobian}
\alias{getJacobian}
\title{Jacobian matrix with interaction strengths}
\usage{
getJacobian(FM, BM, AE, GE, dead, externals, diagonal = 0)
}
\arguments{
\item{FM}{A square flowmatrix, source compartments as rows,
sink compartments as columns.}

\item{BM}{Numeric vector with biomasses of all compartments, must be in the same order as the flow matrix.}

\item{AE}{Numeric vector with assimilation efficiencies of all
compartments,must be in the same order as the flow matrix.
AE should be set to NA for non-faunal compartments.}

\item{GE}{Numeric vector with growth efficiencies of all compartments,
must be in the same order as the flow matrix. GE should be set to NA
for non-faunal compartments.}

\item{dead}{Character vector with all names of detritus and nutrient
compartments (everything that is not fauna).}

\item{externals}{Character vector with all names of external
compartments, i.e. which have no biomass.}

\item{diagonal}{Either a single value, a numeric vector, or the
charcter string "model". A single value with set all diagonal
values to this number, a vector will set the diagonal to this
user-specified diagonal, "model" will calculate diagonal values from
the energy-flux model. Default is a zero diagonal.}
}
\value{
A matrix containing interaction strengths, i.e. the
effect of the resources (rows) on the consumers (columns) - for all
interactions in the food web.
}
\description{
This functions calculates interaction strengths from a resolved energy-flux
food web model and uses these values as entries for a Jacobian matrix.
The diagonal can be either be user-defined, set to zero (default), or calculated
from the energy-flux model.
}
\examples{
getJacobian(FM = Flowmatrix(lim), BM = lim$Components$val)
}
