% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{getJacobian}
\alias{getJacobian}
\title{Jacobian matrix from a food web model}
\usage{
getJacobian(model = stop("Model input required"))
}
\arguments{
\item{model}{(required) A named list containing elements with food web model data.
One list element named \emph{type} must exist and should either be the string "ODE", "EF", or "LIM".
\itemize{
\item \bold{ODE} should be used if the model is dynamic and comprises a set of
ordinary differential equations. Other list element required for ODE type models are
\emph{func}, \emph{y}, and \emph{parms} to be used by the \code{\link{getJacobianODE}}
function.
\item \bold{EF} should be used if the model is a quantified energy flux model. Other list elements required
for EF models are \emph{FM}, \emph{BM}, \emph{AE}, and \emph{GE} to be used by the
\code{\link{getJacobianEnergyFlux}} function.
\item \bold{LIM} should be used if the model is a linear inverse model created with the package LIM.
Other list elements required for LIM type models are \emph{lim} containing a read-in LIM, \emph{aTag}
and \emph{gTAg} containing the tag to search for the assimilated and growht part respectively.
Optional are \emph{web} containing the flow solutions.
}}
}
\value{
This function returns a matrix containing interaction strengths, i.e. the
effect of the resources (rows) on the consumers (columns) - for all
interactions in the food web.
}
\description{
This is a wrapper function that reviews the given food web model and redirects
the input to the correct function for obtaining interaction strengths in a Jacobian matrix.
}
\seealso{
\code{\link{getJacobianODE}}, \code{\link{getJacobianEnergyFlux}}.
}
