% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{getJacobian}
\alias{getJacobian}
\title{Jacobian matrix from a food web model}
\usage{
getJacobian(model = stop("Model input required"))
}
\arguments{
\item{model}{(required) A named list containing elements with food web model data.
One list element named \code{type} denoting the type of input model must exist
and should either be the string "ODE", "EF", or "LIM".
\itemize{
\item \bold{"ODE"} should be used if the model is a set of
ordinary differential equations. Other list element required for ODE type models are
\code{func} and \code{y}. The element \code{parms} is optional. All data will be used
in the \code{\link{getJacobianODE}} function which relies on the rootSolve package.
\item \bold{"EF"} should be used if the model is a quantified energy flux model. Other list elements required
for EF models are \code{FM}, \code{BM}, \code{AE}, and \code{GE} to be used by the
\code{\link{getJacobianEnergyFlux}} function.
\item \bold{"LIM"} should be used if the model is a linear inverse model created
with the package LIM and is formatted to contain tags.
The other list element required for LIM type models is \code{LIM} containing a read-in LIM.
If the LIM is resolved the flow solutions can be provided in the list element \code{web}.
If the LIM is not resolved, the function will use the parsimonious (least distance) solution.
}}
}
\value{
This function returns a matrix containing interaction strengths, i.e. the
effect of the resources (rows) on the consumers (columns) - for all
interactions in the food web.
}
\description{
This is a wrapper function that reviews the given food web model and redirects
the input to the correct function for obtaining interaction strengths in a Jacobian matrix.
}
\section{ODE models}{

A ODE type model should be set-up in such a way that it can be used by the \code{rootSolve} package.
For details, please refer to the documentation of this package.
\itemize{
\item{The element \code{type} should be "ODE".}
\item{The element \code{func} should contain a function that calculates the rate of change for all
compartments i.e. a set of ODEs.}
\item{The element \code{y} is a named vector with the initial state (i.e. biomass) of compartments,
which is needed by the function in \code{func}.
In this respect is the data in \emph{y} similar to the BM vector that needs to be provided
for the EF type models.}
\item{The element \code{parms} is optional and contain parameters that can be used by the function
in \code{func}.}
}
}

\section{Energy Flux models}{

Required elements, optional elements.
Effect of diagonal settings.
Concept and References.
Please see documentation of this function.
}

\section{Linear Inverse models}{

Required elements: Read(LIM) as input. optional Setup(readLIM).
web solutions. How to get these (least distance or likelihood)
Effect of diagonal see EF.
How to use the tags. Model must be set up in a specific way.
}

\seealso{
\code{\link{getJacobianODE}}, \code{\link{getJacobianEnergyFlux}},
\code{\link{extractLIMdata}}.
}
