% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jacobian.R
\name{getJacobianODE}
\alias{getJacobianODE}
\title{Jacobian matrix from a ODE model}
\usage{
getJacobianODE(y, func, parms)
}
\description{
This function produces a Jacobian matrix from a model defined as a set of
ordinary differential equations.
}
\seealso{
The package rootSolve for full documentation on \code{\link[rootSolve]{jacobian.full}}.
}
