% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getMR}
\alias{getMR}
\title{Get mortality rates from a LIM.}
\usage{
getMR(BM, web, vars, mTag = NULL, verbose = T)
}
\arguments{
\item{BM}{(required) A named numeric vector containing biomasses of all LIM components.}

\item{web}{(required) A named numeric vector with the flow solutions.}

\item{vars}{(required) A named numeric vector with the LIM variables.}

\item{mTag}{(optional) Tag assigned to the flows or variables containing the natural, i.e.
non-predatory, mortality. Default is set to "mort". Not case sensitive.}

\item{verbose}{(optional) Should the function provide messages?}
}
\value{
Returns a named vector with mortality rates (per unit time).
}
\description{
This function calculates the mortality rates (per unit time) of all organisms from a LIM.
}
\details{
In order for this function to work, the LIM must be set-up in a specific way.
The flows or variables representing total natural mortality of an organism must be named as:
"compartmentTag" or "tagCompartment" (not case sensitive).
For more information on setting up the LIM, please review the vignette.
}
\references{
\code{\link{LIM}} package, Soetaert & van Oevelen 2015.
}
\seealso{
\code{getVariables}
}
