% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lim_data.R
\name{getVariables}
\alias{getVariables}
\title{Calculate value of LIM variables from flow solutions.}
\usage{
getVariables(readLIM, web = NULL, verbose = T)
}
\arguments{
\item{readLIM}{(required) A linear inverse model read into the R environment with the R-package LIM.
Can be achieved with \code{readLIM <- Read("path_to_input_file")}}

\item{web}{(optional) The solved (food) web problem, i.e. the values of the unknowns;
if not specified, the model is solved first, using Ldei.}
}
\value{
Returns a named vector with all variables.
}
\description{
This function calculates the value of variables as defined in the LIM from the flow solutions.
}
\details{
Variables in the LIM can be defined as the sum of flows, other variables and parameters.
This function calculates the value of these variables based on the flow solutions and LIM parameters.
}
\references{
\code{\link{LIM}} package, Soetaert & van Oevelen 2015.
}
